refactor(cli): move run command to app/cli/run.py
Cuts run_cmd out of app/main.py into app/cli/run.py, which imports
console + _parse_inputs from cli._shared and the rendering helpers
from cli.render. Removes _parse_inputs re-import from app/main.py
since no remaining command body uses it.
